restaurant/requirements.txt
asgiref==3.7.2
Django==4.2.7
djangorestframework==3.14.0
factory-boy==3.3.0
Faker==22.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
